# Improve performance around ignoreCase