@johnynek 2020-11-10
related to #5